requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
python-dotenv==0.14.0
mysqlclient==2.0.1
flask-login==0.5.0
flask-wtf==0.14.3
email-validator==1.1.1